+++ python/GafferOSLUITest/OSLShaderUITest.py
@@ -36,6 +36,8 @@
 
 import pathlib
 
+import imath
+
 import Gaffer
 import GafferUI
 
@@ -109,6 +111,49 @@ class OSLShaderUITest( GafferOSLTest.OSLTestCase ) :
 		)
 		Gaffer.NodeAlgo.applyUserDefaults( s )
 		self.assertEqual( s["parameters"]["spline"]["interpolation"].getValue(), Gaffer.SplineDefinitionInterpolation.MonotoneCubic )
+
+	def testActivatorMetadata( self ) :
+
+		s = self.compileShader( pathlib.Path( __file__ ).parents[1] / "GafferOSLTest" / "shaders" / "activatorMetadata.osl" )
+		n = GafferOSL.OSLShader()
+		n.loadShader( s )
+
+		parameters = n["parameters"]
+		for i in range( 5 ):
+			parameters["i"].setValue( i )
+			self.assertEqual( Gaffer.Metadata.value( parameters["test1"], "layout:activator" ), i > 2 )
+
+		self.assertEqual( Gaffer.Metadata.value( parameters["test2"], "layout:activator" ), False )
+		parameters["s"].setValue( "foo" )
+		self.assertEqual( Gaffer.Metadata.value( parameters["test2"], "layout:activator" ), True )
+
+		parameters["i"].setValue( 9 )
+		self.assertEqual( Gaffer.Metadata.value( parameters["test3"], "layout:visibilityActivator" ), False )
+		parameters["i2"].setValue( 8 )
+		self.assertEqual( Gaffer.Metadata.value( parameters["test3"], "layout:visibilityActivator" ), True )
+
+		self.assertEqual( Gaffer.Metadata.value( parameters["test4"], "layout:visibilityActivator" ), False )
+		parameters["c"].setValue( imath.Color3f( 0.2, 0.4, 0 ) )
+		self.assertEqual( Gaffer.Metadata.value( parameters["test4"], "layout:visibilityActivator" ), True )
+
+	def testConditionalVisMetadata( self ) :
+
+		shader = self.compileShader( pathlib.Path( __file__ ).parents[1] / "GafferOSLTest" / "shaders" / "conditionalVisMetadata.osl" )
+		node = GafferOSL.OSLShader()
+		node.loadShader( shader )
+
+		parameters = node["parameters"]
+		for i in range( 5 ):
+			parameters["i"].setValue( i )
+			self.assertEqual( Gaffer.Metadata.value( parameters["test1"], "layout:activator" ), i > 2 )
+
+		self.assertEqual( Gaffer.Metadata.value( parameters["test2"], "layout:activator" ), False )
+		parameters["s"].setValue( "foo" )
+		self.assertEqual( Gaffer.Metadata.value( parameters["test2"], "layout:activator" ), True )
+
+		self.assertEqual( Gaffer.Metadata.value( parameters["test3"], "layout:visibilityActivator" ), False )
+		parameters["i2"].setValue( 72 )
+		self.assertEqual( Gaffer.Metadata.value( parameters["test3"], "layout:visibilityActivator" ), True )
 
 	def tearDown( self ) :