+++ python/GafferRenderManUITest/RenderManShaderUITest.py
@@ -42,6 +42,8 @@ from xml.etree import ElementTree
 import IECore
 
 import Gaffer
+import GafferOSL
+import GafferOSLUI
 import GafferUITest
 
 import GafferRenderMan
@@ -152,11 +154,71 @@ class RenderManShaderUITest( GafferUITest.TestCase ) :
 						self.assertNotIn( "{}:".format( parameter.getName() ), description )
 						self.assertEqual( description.count( "<p>" ), description.count( "</p>" ) )
 
+					# Check that there are no errors or warnings when evaluating
+					# conditional visibility.
+					Gaffer.Metadata.value( parameter, "layout:activator" )
+					Gaffer.Metadata.value( parameter, "layout:visibilityActivator" )
+
 				shadersLoaded.add( argsFile.stem )
 
 		# Guard against shaders being moved and this test therefore not
 		# loading anything.
 		self.assertIn( "PxrSurface", shadersLoaded )
+
+	def testSpecificActivator( self ) :
+
+		shader = GafferRenderMan.RenderManShader()
+		shader.loadShader( "PxrSurface" )
+
+		self.assertEqual( Gaffer.Metadata.value( shader["parameters"]["diffuseExponent"], "layout:visibilityActivator" ), True )
+		shader["parameters"]["diffuseRoughness"].setValue( 0.1 )
+		self.assertEqual( Gaffer.Metadata.value( shader["parameters"]["diffuseExponent"], "layout:visibilityActivator" ), False )
+
+		shader.loadShader( "LamaConductor" )
+		self.assertEqual( Gaffer.Metadata.value( shader["parameters"]["fresnelMode"], "layout:activator" ), True )
+
+		self.assertEqual( Gaffer.Metadata.value( shader["parameters"]["exteriorIOR"], "layout:activator" ), False )
+		shader["parameters"]["overrideExteriorIOR"].setValue( True )
+		self.assertEqual( Gaffer.Metadata.value( shader["parameters"]["exteriorIOR"], "layout:activator" ), True )
+
+	def testAllOSLActivators( self ) :
+
+		# Load every RenderMan OSL shader and check that there are no
+		# errors or warnings when evaluating their conditional visibility.
+
+		shaderDir = pathlib.Path( os.environ["RMANTREE"] ) / "lib" / "shaders"
+		shadersLoaded = set()
+
+		for osoFile in shaderDir.glob( "*.oso" ) :
+
+			with self.subTest( shader = osoFile.stem ) :
+
+				node = GafferOSL.OSLShader( osoFile.stem )
+				node.loadShader( osoFile.stem )
+
+				for parameter in node["parameters"] :
+					Gaffer.Metadata.value( parameter, "layout:activator" )
+					Gaffer.Metadata.value( parameter, "layout:visibilityActivator" )
+
+				shadersLoaded.add( osoFile.stem )
+
+		# Guard against shaders being moved and this test therefore not
+		# loading anything.
+		self.assertIn( "PxrHairColor", shadersLoaded )
+
+	def testSpecificOSLActivators( self ) :
+
+		shader = GafferOSL.OSLShader()
+		shader.loadShader( "PxrHairColor" )
+
+		self.assertEqual( Gaffer.Metadata.value( shader["parameters"]["melanin"], "layout:activator" ), True )
+
+		self.assertEqual( Gaffer.Metadata.value( shader["parameters"]["melanin"], "layout:visibilityActivator" ), True )
+		self.assertEqual( Gaffer.Metadata.value( shader["parameters"]["Color"], "layout:visibilityActivator" ), False )
+
+		shader["parameters"]["mode"].setValue( 1 ) # Artistic
+		self.assertEqual( Gaffer.Metadata.value( shader["parameters"]["melanin"], "layout:visibilityActivator" ), False )
+		self.assertEqual( Gaffer.Metadata.value( shader["parameters"]["Color"], "layout:visibilityActivator" ), True )
 
 if __name__ == "__main__":
 	unittest.main()