+++ src/GafferOSL/OSLShader.cpp
@@ -475,7 +475,10 @@ Plug *loadStringArrayParameter( const OSLQuery::Parameter *parameter, const Inte
 		defaultValueDataWritable.resize( parameter->sdefault.size() );
 		for( size_t i = 0; i < parameter->sdefault.size(); i++ )
 		{
-			defaultValueDataWritable[i] = parameter->sdefault[i].c_str();
+			if( !parameter->sdefault[i].empty() )
+			{
+				defaultValueDataWritable[i] = parameter->sdefault[i].c_str();
+			}
 		}
 	}
 
@@ -1509,6 +1512,20 @@ void OSLShader::reloadShader()
 	queryCache().erase( namePlug()->getValue() );
 	g_metadataCache.erase( namePlug()->getValue() );
 	Shader::reloadShader();
+}
+
+namespace GafferOSL
+{
+
+// Forward declare function defined in OSLExpressionEngine.cpp. It happens to be easier to implement
+// activator expressions in there, but in terms of public API they are more at home on OSLShader.
+bool evaluateActivatorExpression( const std::string &expression, const Gaffer::Plug *parameterPlug, const Gaffer::Context *context );
+
+} // namespace GafferOSL
+
+bool OSLShader::evaluateActivatorExpression( const std::string &expression ) const
+{
+	return GafferOSL::evaluateActivatorExpression( expression, parametersPlug(), Context::current() );
 }
 
 bool OSLShader::registerCompatibleShader( const IECore::InternedString shaderType )